vfio:ubus: add vfio ub cfg0 and cfg1 basic attr perm
driver inclusion
category: bugfix
bugzilla: https://atomgit.com/openeuler/kernel/issues/6695
CVE: NA

(cherry-pick from OLK-6.6 commit a9067e87b155)

-----------------------------------------------------------

The VFIO UB needs to add multiple configuration space
registers to adapt to protocol changes.

Fixes: 61a1ca5cb417 ("vfio:ubus: Init vfio ub config space 0 perm bits during module init")

+++ drivers/vfio/ubus/vfio_ub_intrs.c
@@ -145,7 +145,7 @@ static int vfio_ub_intr_set_block(struct vfio_ub_core_device *vdev, unsigned int
 	if ((start >= vdev->num_ctx) || ((start + count) > vdev->num_ctx))
 		return -EINVAL;
 
-	for (i = 0, j = start; i < count; i++, j++) {
+	for (i = 0, j = (int)start; i < (int)count; i++, j++) {
 		int fd = fds ? fds[i] : -1;
 
 		ret = vfio_ub_intr_set_vector_signal(vdev, j, fd);
